Release new version with Permission model and migration changes

+++ database/migrations/2025_04_27_180100_create_permission_role_table.php
@@ -11,14 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('permission_role', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('permission_id')->constrained()->onDelete('cascade');
-            $table->foreignId('role_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-            
-            $table->unique(['permission_id', 'role_id']);
-        });
+        if (!Schema::hasTable('permission_role')) {
+            Schema::create('permission_role', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('permission_id')->constrained()->onDelete('cascade');
+                $table->foreignId('role_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+                
+                $table->unique(['permission_id', 'role_id']);
+            });
+        }
     }
 
     /**
